requirements.txt
tensorflow-cpu==2.12.0
keras==2.12.0
numpy
pillow
Flask==2.2.5
scikit-image
tqdm
h5py